+++ ObjectPrinting/Config.cs
@@ -2,58 +2,44 @@
 using System.Collections.Generic;
 using System.Reflection;
 
-namespace ObjectPrinting.Solved
+namespace ObjectPrinting
 {
     public class Config
     {
-        public List<Type> excludedTypes;
-        public List<PropertyInfo> exludedFields;
-        public Dictionary<PropertyInfo, Delegate> fieldSerializers;
-        public Dictionary<Type, Delegate> typesSerializer;
-
-        public Config(Config config)
-        {
-            excludedTypes = new List<Type>();
-            typesSerializer = new Dictionary<Type, Delegate>();
-            exludedFields = new List<PropertyInfo>();
-            fieldSerializers = new Dictionary<PropertyInfo, Delegate>();
-        }
+        public readonly List<Type> ExcludedTypes;
+        public readonly List<PropertyInfo> ExludedFields;
+        public readonly Dictionary<PropertyInfo, Delegate> FieldSerializers;
+        public readonly Dictionary<Type, Delegate> TypesSerializer;
 
         public Config()
         {
-            excludedTypes = new List<Type>();
-            typesSerializer = new Dictionary<Type, Delegate>();
-            exludedFields = new List<PropertyInfo>();
-            fieldSerializers = new Dictionary<PropertyInfo, Delegate>();
+            ExcludedTypes = new List<Type>();
+            TypesSerializer = new Dictionary<Type, Delegate>();
+            ExludedFields = new List<PropertyInfo>();
+            FieldSerializers = new Dictionary<PropertyInfo, Delegate>();
         }
 
         public bool IsExcluded(PropertyInfo propertyInfo)
         {
-            return exludedFields.Contains(propertyInfo) || excludedTypes.Contains(propertyInfo.PropertyType);
+            return ExludedFields.Contains(propertyInfo) || ExcludedTypes.Contains(propertyInfo.PropertyType);
         }
 
         public bool IsSpecialSerialize(PropertyInfo propertyInfo, object element, out string result)
         {
             result = "";
-            if (!fieldSerializers.ContainsKey(propertyInfo))
+            if (!FieldSerializers.ContainsKey(propertyInfo))
                 return false;
-            result = fieldSerializers[propertyInfo].DynamicInvoke(element)?.ToString();
+            result = FieldSerializers[propertyInfo].DynamicInvoke(element)?.ToString();
             return true;
         }
 
         public bool IsSpecialSerialize(Type type, object element, out string result)
         {
             result = "";
-            if (!typesSerializer.ContainsKey(type))
+            if (!TypesSerializer.ContainsKey(type))
                 return false;
-            result = typesSerializer[type].DynamicInvoke(element)?.ToString();
+            result = TypesSerializer[type].DynamicInvoke(element)?.ToString();
             return true;
-        }
-
-        public Config Exclude(Type type)
-        {
-            excludedTypes.Add(type);
-            return new Config(this);
         }
     }
 }
